requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
onvif-zeep==0.2.12
Werkzeug==3.0.1
python-dotenv==1.0.0
requests==2.31.0
Pillow>=10.2.0
opencv-python>=4.9.0
gunicorn==21.2.0
